# openapi.yaml
openapi: 3.1.0
info:
  title: Go System API
  version: 1.0.0
  description: RESTful API for Linux system management
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
paths:
  /auth/login:
    post:
      summary: Authenticate user
      description: |
        Authenticate a user and receive a JWT token.
        
        **Authentication Methods:**
        1. Built-in admin account (development builds only):
           - Username: admin
           - Password: admin123
           - Note: Disabled in production builds for security
        
        2. Linux system users (recommended for production):
           - Any valid Linux user with their system credentials
           - Uses system authentication via 'su' command
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /network/interfaces:
    get:
      summary: List network interfaces
      tags:
        - Network
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of interfaces
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterfacesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /network/interfaces/{name}/up:
    put:
      summary: Bring interface up
      tags:
        - Network
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Interface brought up
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Interface not found
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /network/interfaces/{name}/down:
    put:
      summary: Bring interface down
      tags:
        - Network
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Interface brought down
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Interface not found
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /network/interfaces/{name}/address:
    post:
      summary: Configure IP address
      tags:
        - Network
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
      responses:
        '200':
          description: Address configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Interface not found
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  
    put:
      summary: Update IP address
      tags:
        - Network
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
      responses:
        '200':
          description: Address updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Interface not found
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  
    delete:
      summary: Delete IP address
      tags:
        - Network
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: address
          in: query
          required: true
          schema:
            type: string
            format: ipv4
      responses:
        '200':
          description: Address deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Interface or address not found
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /network/bridges:
    get:
      summary: List network bridges
      tags:
        - Network
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of bridges
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      bridges:
                        type: array
                        items:
                          $ref: '#/components/schemas/Interface'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /network/bridge:
    post:
      summary: Create network bridge
      tags:
        - Network
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BridgeRequest'
      responses:
        '200':
          description: Bridge created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '409':
          description: Bridge already exists
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /network/bridge/{name}:
    put:
      summary: Update network bridge
      tags:
        - Network
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BridgeRequest'
      responses:
        '200':
          description: Bridge updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Bridge not found
        '400':
          description: Not a bridge interface
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      summary: Delete network bridge
      tags:
        - Network
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bridge deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Bridge not found
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /network/bonds:
    get:
      summary: List network bonds
      tags:
        - Network
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of bonds
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      bonds:
                        type: array
                        items:
                          $ref: '#/components/schemas/Interface'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /network/bond:
    post:
      summary: Create network bond
      tags:
        - Network
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BondRequest'
      responses:
        '200':
          description: Bond created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '409':
          description: Bond already exists
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /network/bond/{name}:
    put:
      summary: Update network bond
      tags:
        - Network
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BondRequest'
      responses:
        '200':
          description: Bond updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Bond not found
        '400':
          description: Not a bond interface
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      summary: Delete network bond
      tags:
        - Network
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bond deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Bond not found
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /network/vlans:
    get:
      summary: List VLAN interfaces
      tags:
        - Network
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of VLANs
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      vlans:
                        type: array
                        items:
                          $ref: '#/components/schemas/Interface'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /network/vlan:
    post:
      summary: Create VLAN interface
      tags:
        - Network
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VLANRequest'
      responses:
        '200':
          description: VLAN created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '409':
          description: VLAN already exists
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /network/vlan/{name}:
    put:
      summary: Update VLAN interface
      tags:
        - Network
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VLANRequest'
      responses:
        '200':
          description: VLAN updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: VLAN not found
        '400':
          description: Not a VLAN interface
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      summary: Delete VLAN interface
      tags:
        - Network
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: VLAN deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: VLAN not found
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /storage/disks:
    get:
      summary: List storage disks
      tags:
        - Storage
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of disks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisksResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /storage/mount:
    post:
      summary: Mount filesystem
      tags:
        - Storage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MountRequest'
      responses:
        '200':
          description: Filesystem mounted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Device not found
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /storage/unmount:
    post:
      summary: Unmount filesystem
      tags:
        - Storage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnmountRequest'
      responses:
        '200':
          description: Filesystem unmounted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /storage/format:
    post:
      summary: Format disk
      tags:
        - Storage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormatRequest'
      responses:
        '200':
          description: Disk formatted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Device not found
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /storage/lvm/vgs:
    get:
      summary: List LVM volume groups
      tags:
        - Storage
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Volume groups listed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VolumeGroupsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /storage/lvm/lvs:
    get:
      summary: List LVM logical volumes
      tags:
        - Storage
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logical volumes listed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogicalVolumesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /storage/lvm/pvs:
    get:
      summary: List LVM physical volumes
      tags:
        - Storage
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Physical volumes listed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhysicalVolumesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /storage/lvm/vg:
    post:
      summary: Create LVM volume group
      tags:
        - Storage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVGRequest'
      responses:
        '200':
          description: Volume group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /storage/lvm/lv:
    post:
      summary: Create LVM logical volume
      tags:
        - Storage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLVRequest'
      responses:
        '200':
          description: Logical volume created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /storage/iscsi/discover:
    post:
      summary: Discover iSCSI targets
      tags:
        - Storage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ISCSIDiscoverRequest'
      responses:
        '200':
          description: Targets discovered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ISCSITargetsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /storage/iscsi/sessions:
    get:
      summary: List iSCSI sessions
      tags:
        - Storage
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sessions listed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ISCSISessionsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          properties:
            status:
              enum: [error]
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: |
            Username for authentication.
            - For development builds: 'admin' with password 'admin123' is available
            - For production builds: Only Linux system users are allowed
        password:
          type: string
          description: User password
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                token:
                  type: string
                expires_at:
                  type: integer
...
    InterfacesResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                interfaces:
                  type: array
                  items:
                    $ref: '#/components/schemas/Interface'
...